Include security deposit information on Lease API

+++ backend/api/Areas/Leases/Models/Lease/SecurityDepositModel.cs
@@ -58,6 +58,6 @@ namespace Pims.Api.Areas.Lease.Models.Lease
         /// <summary>
         /// get/set - The annual interest rate.
         /// </summary>
-        public int AnnualInterestRate { get; set; }
+        public decimal AnnualInterestRate { get; set; }
     }
 }
